Backend/requirements.txt
Flask>=3.0.0
flask-restx>=1.3.0
python-dotenv==1.0.0
Flask-Cors>=3.0.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended>=4.6.0
bcrypt>=4.0.0
cloud-sql-python-connector[pymysql]
openai>=1.0.0
requests>=2.31.0
